requirements.txt
appnope==0.1.0
astroid==1.6.0
Babel==2.5.1
backports.functools-lru-cache==1.4
backports.shutil-get-terminal-size==1.0.0
certifi==2017.11.5
chardet==3.0.4
configparser==3.5.0
decorator==4.1.2
Django==1.11.7
django-nose==1.4.5
django-phonenumber-field==1.3.0
djangorestframework==3.7.3
enum34==1.1.6
Faker==0.8.8
httpie==0.9.9
idna==2.6
ipaddress==1.0.19
ipdbplugin==1.4.5
ipython==5.5.0
ipython-genutils==0.2.0
isort==4.2.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
nose==1.3.7
pathlib2==2.3.0
pexpect==4.3.1
phonenumberslite==8.8.5
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
pylint==1.8.1
python-dateutil==2.6.1
pytz==2017.3
requests==2.18.4
scandir==1.6
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
text-unidecode==1.1
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
wrapt==1.10.11
